Better JWT
manageJWT acts as a basic middleware for the API's to easily deal with JWT.

+++ api/history.js
@@ -1,11 +1,18 @@
-const { bcrypt, crypto, History, jwt, JWT_SECRET, mongoose, express } = require('../modules');
+const {History, express } = require('../modules');
 const router = express.Router();
-//TO-DO: remove unneeded dependencies above when finished ^^^
+const manageJWT = require('../manageJWT');
 
-//TO-DO: JWT
+//TO-DO: Send better messages upon success
 router.post('/api/history-new', async (req, res) => {
-    const { userID, restaurantID, liked } = req.body;
+    const { token, restaurantID, liked } = req.body;
     try {
+
+        //JWT verification
+        let userID = manageJWT.returnUser(token);
+        if (userID == null)
+        {
+            return res.status(401).send('Token unverified or expired');
+        }
 
         // Check if a user has a history already
         const existingHistory = await History.findOne({ userID : userID, restaurantID : restaurantID });
@@ -28,45 +35,69 @@ router.post('/api/history-new', async (req, res) => {
         });
         await newHistory.save();
 
-        // Return success
-        res.status(200).send('History created successfully');
+        let newToken = manageJWT.refresh( token );
+
+        // History created successfully
+        res.status(200).send({ token : newToken });
     } catch (err) {
         console.error(err);
         res.status(500).send('Internal server error');
     }
 });
 
-//TO-DO: JWT
+//TO-DO: Send better messages upon success
 router.post('/api/history-visted', async (req, res) => {
-    const { userID, restaurantID} = req.body;
+    const { token, restaurantID } = req.body;
     try {
-        // Check if a user with the same email address already exists
-        const existingHistory = await History.findOne({ userID : userID, restaurantID : restaurantID });
-        if (existingHistory) {
-            existingHistory.isVisited = true;
-            existingHistory.save()
-
-            return res.status(200).send('History updated');
-        }
-        else
+        
+        //JWT verification
+        let userID = manageJWT.returnUser(token);
+        if (userID == null)
         {
-            return res.status(404).send('History object does not exist');
+            return res.status(401).send('Token unverified or expired');
         }
-    } catch (err) {
+    
+        //Find a user
+        const visitedHistory = await History.findOne({ userID : userID, restaurantID : restaurantID });
+        if (visitedHistory){
+            visitedHistory.isVisited = true;
+            visitedHistory.save();
+
+            let newToken = manageJWT.refresh(token);
+
+            //History visted updated
+            return res.status(200).send({ token: newToken });
+        }
+        else{
+
+            return res.status(404).send('No History found') ;
+        }
+    }catch (err) {
         console.error(err);
         res.status(500).send('Internal server error');
     }
 });
 
-//TO-DO: JWT
+//TO-DO: Send better messages upon success
 router.post('/api/history-delete', async (req, res) => {
-    const { userID, restaurantID } = req.body;
+    const { token, restaurantID } = req.body;
     try {
+        
+        //JWT verification
+        let userID = manageJWT.returnUser(token);
+        if (userID == null)
+        {
+            return res.status(401).send('Token unverified or expired');
+        }
+
         //Deletes user if they exist, returns the deleted object
         const deletedHistory = await History.findOneAndDelete({ userID : userID, restaurantID : restaurantID });
             if (deletedHistory){
 
-                return res.status(200).send('History deleted');
+                let newToken = manageJWT.refresh(token);
+
+                //History deleted
+                return res.status(200).send({ token: newToken });
             }
             else{
 
@@ -79,13 +110,26 @@ router.post('/api/history-delete', async (req, res) => {
     }
 });
 
-//TO-DO: JWT, and more consise wishlist? (right now submitting every history element unabashedly)
+//TO-DO: Send better messages upon success
+//TO-DO: More consise wishlist? (right now submitting every history element unabashedly)
 //Also, maybe return 404 error if no items in list???
 router.post('/api/wishlist', async (req, res) => {
-    const { userID } = req.body;
+    const { token } = req.body;
     try {
+        
+        //JWT verification
+        let userID = manageJWT.returnUser(token);
+        if (userID == null)
+        {
+            return res.status(401).send('Token unverified or expired');
+        }
+
         const wishlist = await History.find({ userID : userID, liked : true, isVisited : false});
-        res.json({ wishlist });
+
+        let newToken = manageJWT.refresh(token);
+
+        //return wishlist
+        res.json({ token : newToken, wishlist });
 
     } catch (err) {
         console.error(err);
